Add a `component-model` feature to wat/wast/wasm-encoder (#1844)
* Add a crate feature for component-model support to wast/wat

In case folks depend on these crates but aren't interested in the
component support this enables compiling it all out.

* Add a Cargo feature for components in wasm-encoder

If users are interested it can be used to reduce the compile-time of the
crate and amount of code brought in by removing much of it at compile
time. Support remains enabled by default, however.

* Fix standalone build of wasm-smith

* Fix standalone build of wasm-metadata