Add a `component-model` feature to wasmparser  (#1845)
* Split-out component-related types in `wasmparser`

This commit splits the `types.rs` in `wasmparser` into a `types.rs` and
`component_types.rs`. The goal here is to split all component-related
logic to a separate file instead of having everything bundled together.
This will make it easier to conditionally include component-related bits
in the future and additionally make each module more understandable by
focusing only on one piece of functionality.

* Make `component_types` a part of `wasmparser`'s API

Don't `pub use` all the types from that module into the preexisting
`types` module. This helps reduce the cognitive load when looking at the
`types` module by keeping all the component bits in one place.

* Add a `component-model` feature to `wasmparser`

This commit gates the implementation of the component-model proposal in
`wasmparser` behind a Cargo feature named `component-model`. This is
intended to lower the binary size and compile time of this crate in
situations where components are not needed. Additionally this helps
factor logic cleanly between core module and component layers both at
the API and internal levels.

* Add some more checks to CI

* Fix fuzzer and bench build

* Review feedback