Fix generating shims with duplicate names
This fixes an issue with the semver-merging logic of `wit-component`
where during the componentization process due to
insertion-with-replacement behavior the same name could be assigned to
two different shims (as shim names are based on the map length which
wouldn't change if a replacement was inserted). The fix in this commit
is to avoid replacement and prefer the original name when a shim is inserted.

Closes #1850