`wast`: Add `#[non_exhaustive]` to `WastArg` and `WastRet` (#1891)
* add #[non_exhaustive] to WastArg and WastRet

This attribute is required because depending on #[cfg(feature = "component-model")] WastArg and WastRet can be enums with a single variant. Since crate features are resolved together this might cause problems in a crate due to feature usage of another crate or dependency, e.g. in the same workspace, with respect to irrefutable patterns.

* fix errors after attribute additions

* remove now outdated comment