Prepare to update spec test suite (#1909)
* Prepare to update spec test suite

Some minor updates in preparation for WebAssembly/testsuite#110:

* Don't limit the size of tables in the validator and instead defer such
  maximal validation to runtimes themselves.
* Parse the `(pagesize N)` syntax on "inline memories"
* Always parse table/memory limits as 64-bit integers
* Stop matching spec test suite error messages

This last point is something I've tried to hold off on doing for a long
time but as the giant `error_matches` function continues to grow and
become more unwieldy it has become less and less tenable to do this.
Overall it doesn't seem too beneficial to keep maintaining this
especially in the face of numerous proposals, so instead basically
remove it entirely.

* Rename test