Remove quadratic behavior cloning rec groups (#1952)
The call to `clone_rec_group` would create a new vector each time it
was called of all clonable rec groups where if an empty rec group is
cloned this creates quadratic behavior where each time a vector is
created and one more item is appended.

The fix in this commit is to avoid creating a precise set of candidates
to clone and just throwing out candidates that aren't applicable.