smith: Fix types recorded in `ref.func`
This commit fixes a minor issue in `wasm-smith` when generating the
`ref.func` instruction when the GC proposal is disabled but
shared-everything-threads is enabled. In this situation all `ref.func`
instructions were recorded as `RefType::FUNCREF` but this wasn't correct
for shared function types. The logic here is applicable both with and
without the GC proposal is enabled so the precise type is always pushed
now.