Explicitly disallow 64-bit/shared memories/tables in components (#1970)
These proposals are not specified how they work with the canonical ABI
just yet. Previously the proposals were not enabled by default so their
off-by-default status largely gated their usage in components but with
memory64 now being on-by-default it's possible to have components using
64-bit linear memories. More care will be needed to update components
and tooling for 64-bit linear memories so for now an error is added to
reject it saying that support is not added yet.