Fix panic processing invalid WIT (#1991)
* Fix panic processing invalid WIT

When an error was being generated part of the error message generation
was producing a panic by accessing some data within `Resolve` before it
was updated. The fix here is to move the data-updating before the error
message generation to resolve the panic.

* Reformat some long lines

Let `rustfmt` be able to wrap the lines

* Fix `cargo test -p wit-parser` in isolation

* Thread `Span` information for stability errors instead