Fix handling stability of `use` of "foreign" types (#2046)
This commit addresses an issue where stability attributes on a `use`
didn't quite work as expected when a type to another package was
referred to. This fix was to update the "merging" process to skip types
being processed in one more location which involved threading some more
contexts around. Additionally `use` items, when elaborated, now contain
their stability instead of the default stability to ensure that's
propagated correctly as well.

cc #1995 but doesn't fix it