Remove support for 33+ flags and multiple returns (#2050)
This commit removes the feature gates of supporting 33-or-more flags in
the component model and also being able to support named or multiple
return types. These have been gated to return an error for some time now
and no issues have arisen so now seems as good a time as any to remove
the internal implementation details. This simplifies the representation
of function return types and is intended to help make the internals
here easier to understand for newcomers.