Refactor how `error-context` is modeled
This commit refactors how the `error-context` type in the async
component model implementation is modeled within these crates.
Previously it was listed as a "defined type" but the binary encoding and
component model itself has it listed as a "primitive type". While mostly
the same the distinction is subtle enough (e.g. around naming rules)
that this commit refactors the wasm-tools crates to match the component
model specification.