Refactor some wit-component internals related to async component model bits (#2062)
* Refactor `Shims::append_indirect`

Pull some code out of this large function into separate functions to
also help simplify the control flow in the loop. The original semantics
of fall-through-for-wit-imports made more sense when there was just 1-2
intrinsics but now that there are many more the control flow here looks
quite odd.

* Simplify shim encoding for error-context intrinsics

* Instead of having a placeholder "empty" options use the actual options
  required for this shim.
* Simplify a `match` on the actual generation of the intrinsics a bit.

* Refactor/simplify `materialize_imports`

Shuffle some things around in to new helper methods to help reduce the
size of this relatively large function.

* Refactor/simplify ABI options for stream/future read/write

This commit refactors the calculation of ABI options for various
intrinsics to have them on-hand earlier to store within each `Shim`.
While here various bits and pieces are refactored to be a bit more
"modern" given the current shape of components.