Refactor how `error-context.*` intrinsics are imported
This commit refactors the `wit-component` conventions used to import
`error-context.*` intrinsics. The first change is to drop the `realloc`
option on the `error-context.debug-message` intrinsic in favor of
inferring that from the "general purpose import realloc" option. This
avoids the need to name a symbol and enables generally relying on there
being a single option for this.

The second change is to change how `encoding=` is specified and bake it
directly into the name instead of having a url-scheme-like encoding with
multiple options. This is in theory the only option necessary for these
intrinsics so it's possible to have a static set of a few names that
might be chosen.