Refactor/improve parsing of component canonical text format
This commit is an improvement and refactoring to the parsing of `(canon
...)` forms in the text format for components. Previously there was a
large amount of duplication to handle the short-hand form of:

```
(core func (canon ...))
```

and the long-hand form of

```
(canon ... (core func))
```

This commit removes this duplication and in the process fixes support
for the long-hand format of `(canon ...)`. This is not heavily used in
tooling of this workspace but it's mentioned in the specification and
should likely be supported regardless.