Promote an `assert!` to a `bail!`
This commit fixes a runtime assertion tripping to instead being a
first-class error returned by `bail!`. This cannot currently be
triggered from the CLI and is only reachable through API usage of the
`wit_parser::Resolve` type. This usage is reachable through generators
such as `wit_bindgen::generate!`, though.

The error here happens when a package is re-added to a `Resolve` twice.
This currently isn't supported and would require some large refactoring
to support. This should probably be fixed at some point in the future to
actually be supported but until that happens it's best to have a
first-class error for this case instead of an internal assertion
tripping.

Closes #1996