wasmparser: add a hint to out-of-bounds type indices errors (#2094)
When an out-of-bounds type index would be in bounds if the GC proposal was
enabled, but it is not, add some extra verbiage to the error message explaining
that. This is useful for people who are playing around with Wasm GC in Wasmtime,
but forget to enable it on the CLI or in their config.

I've taken care to only *add* to the error message, so that the substring
matching we use for `.wast` tests will continue to work.