wit-component: Eagerly define all exported types (#2091)
* wit-component: Eagerly define all exported types

This commit fixes an assertion failure found via the fuzzing of
wit-component. The basic problem is that the `task.return` intrinsic,
and other async intrinsics, need to refer to asynchronous types to be
declared, but while defining some types can be done lazily it's not
always easy to do so. Currently all types are defined lazily as the
intrinsic is declared, but this panics when referring to types defined
in other interfaces as the current type-defining machinery (rightfully)
panics in this situation.

The fix applied in this commit is to update the part of encoding which
eagerly defines type information for imported intrinsics, previously
just resources, to now also define all types for exported interfaces.
This ensures that all defined types necessary will already be available
by the time that intrinsics are defined. Overall `wit-component` is
definitely in need of cleaning up how types are defined, but for now
that's left for a future refactoring...

Otherwise some test expectations are also changing here in this commit
as types were shuffled around and a few dead types were also generated.

* Remove debugging statement