Migrate spec tests to `tests/cli/*`
Similarly to the previous commit of moving towards `wasm-tools wast`
this commit migrates the spec test infrastructure to using this as well.
Spec tests are now run by:

* A script, `ci/generate-spec-tests.rs`, generates files in
  `tests/cli/spec/*.wast` to run each spec test.
* Each test lists the features needed to be enabled for the test to run.
* The tests are then naturally run during the `--test cli` test suite run.

New automation is added to ensure that the version in-tree is up-to-date
to ensure we don't creep in too many files by accident.