Add a `wast` subcommand to the CLI (#2096)
* Add a `wast` subcommand to the CLI

This commit migrates the `tests/roundtrip.rs` file to a `wast`
subcommand of the CLI. This does not actually run tests in terms of
executing WebAssembly but it works as a way to parse the test and
validate, so everything short of runtime execution and related errors.

* Migrate `tests/local/*` to `tests/cli/*`

With the addition of the `wast` subcommand in the previous commit
there's no longer any need to have `tests/local/*`. Additionally it's
now much easier to perform per-test configuration since the flags for
the test are in the test itself.

* Migrate spec tests to `tests/cli/*`

Similarly to the previous commit of moving towards `wasm-tools wast`
this commit migrates the spec test infrastructure to using this as well.
Spec tests are now run by:

* A script, `ci/generate-spec-tests.rs`, generates files in
  `tests/cli/spec/*.wast` to run each spec test.
* Each test lists the features needed to be enabled for the test to run.
* The tests are then naturally run during the `--test cli` test suite run.

New automation is added to ensure that the version in-tree is up-to-date
to ensure we don't creep in too many files by accident.

* Remove the roundtrip test suite

No longer needed!

* Handle some rebase conflicts

* Update submodules in new CI job

* Update testing and contributing docs

* Fix typo in README

* Sync snapshots