threads: add function types to `thread.spawn_indirect`
Initially I implemented `thread.spawn_indirect` without the ability to
check the type of the function to spawn out of an abundance of caution
(see the implementation issues described in [#89]). In the process of
writing out the specification, we convinced ourselves that these
problems should not apply to `thread.spawn_indirect`.

This change adds the function type index necessary for doing some extra
validation of `thread.spawn_indirect` and adds some tests related to
this. One unimplemented TODO is what to do about shared tables:
technically, the table used by a `thread.spawn_indirect` should be
shared but we have so far prevented this in the component model; this
can be resolved later, though.

[#89]: https://github.com/WebAssembly/shared-everything-threads/issues/89