Improve support for memory64/shared-everything-threads in components
Previously the validator in wasm-tools was updated to conservatively
reject all aliases of exports of core memories/tables that were 64-bit
or shared. This is a coarse approximation of the actual rules though
which today disallow using 64-bit memories in the canonical ABI. Instead
this commit updates with a few minor changes to improve the situation
here, namely:

* Aliasing core exports is now allowed for any table and memory type.
* Usage of memories in canonical ABI positions now verifies that the
  memory is of the expected type, using the standard sub-typing.
* Usage of tables in `canon thread.spawn_indirect` now performs a
  type-check via a similar method to ensure the table is of the right type.

This adjusts the behavior of a few tests here and there as well to
account for some slight differences in behavior with these proposals.