Fix/improve validation of components with tags
This commit fixes an issue where the `Tag` kind of core item was allowed
to "leak through" validation of components without the exceptions
proposal active. Additionally creating a core instance with a tag was
consulting the wrong index space and needed fixing too. The end result
is that the `exceptions` proposal is now required for tags to appear
within components.

This so far has been glossed over because any valid component would
require a core module with a tag and those are properly gated, but it's
possible to construct an invalid component with a tag "kind" within it
which doesn't actually have any tags. By forgetting to gate on the
exception-handling proposal it was leaking this "kind" to callers which
weren't necessarily prepared for it.

Finally this also fixes a panic in the text format for components where
named tags still had an `unimplemented!()` for when they were being
registered.