Refactor how liveness of imports is calculated (#2115)
In `wit-component` there's a pass when creating a component to prune
types/functions/etc from imported interfaces if they're not actually
needed by the final component. This logic has not stood well against the
test of time and the loop nowadays looks like: for each import if it's
found within the imports then mark it live. This logic had some special
handling of resource intrinsics but did not take into account any
async/stream/future-related intrinsics.

To fix this issue I've replaced the outdated loop with "for all module
imports add them, as necessary, to the set of live types". This starts
with the source of truth, module imports, and works backwards to WIT
types as opposed to the other way around, starting with WIT types and
testing to see if it's included in the intrinsics list. That fixes the
various cases with async intrinsics and should ensure that all future
intrinsics are also accounted for.

Closes bytecodealliance/wasip3-prototyping#82