Stop passing around `&WasmFeatures` (#2121)
This commit updates the validation phase of wasmparser to avoid passing
around `&WasmFeatures` all over the place. Instead it's now stored
within module/component state in addition to the validator itself to
store a duplicate copy. This should be reasonable to do as
`WasmFeatures` is cheap to clone/copy.

The main motivation for this is that it's a big pain to pass around
these features. Existing code paths often start getting it as a new
argument when new wasm proposals are implemented and it's quite a lot of
threading around to do. There's not really much reason to avoid copies
when everything is a single source of truth from the validator anyway
and the features never change afterwards, so reduce the cognitive
complexity by just storing copies in module/component states.