Return an error earlier on instructions-after-function-end
WebAssembly considers any instructions after the final `end` in a
function to be invalid, and `wasmparser` correctly identifies such
modules as invalid. Currently though if "valid code" is present until
the actual byte end of the function this error is delayed until the
function is finalized. This is currently to avoid the overhead of
checking if we're at the end of the function during the validation of
all operators.

This can have adverse side effects on consumers, however. For example
for the included test case it currently causes `wasmtime compile` to
panic. The reason for this is that compilation continues after the end
of the function and Wasmtime ends up panicking before an error is
returned from wasmparser. The exact reason for the error is a mismatch
in the understanding where Wasmtime thinks a `return` must match the
function's types, but due to the way validation-after-end-of-function
works the validator was validating whatever block was pushed prior.
This mismatch led to a panic in Wasmtime.

The fix here is to prevent the control stack from growing in height
after it has been emptied out. That forces the error to be returned
sooner and also ensures that once the control stack is emptied again
it'll never grow. This continues to avoid a (hopefully unnecessary)
check on all instructions as to whether the function has ended and keeps
the cost localized to just control-frame-modification instructions.