add `subtask.cancel` intrinsic
This has not yet been added to the spec, but we've got a general idea of how it
will look and behave: https://github.com/WebAssembly/component-model/issues/495

Note that, since there's not spec for it yet, I've picked the next unused binary
discriminant in the sequence: 0x24.  That should be easy enough to change if
necessary once the spec is updated.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>