Alternative fix for conflicting `name` and `[method]name.name`
This commit fixes another fuzz bug after #2141 didn't actually fix other
variations of the original bug. The fix here is to ensure that the
creation of `[method]name.name` basically just isn't possible because if
`name` is present, which is required for resources, then
`[method]name.name` is not allowed to be present.