Pre-sort WIT worlds at rest, update printing
This commit moves a sort that previously happened during translation to
wasm to happening during general elaboration instead, meaning that
at-rest worlds are now pre-sorted in this order. Additionally printing
of WIT worlds was updated to manually respect this order where
interfaces are printed first, then types, then functions. As a result of
this change many test expectations have been updated to shuffle items
around.

The motivation for this commit is to fix fuzz failures about
round-tripping WIT documents through wasm. With `include` now supported
more test cases are showing up which stress new paths of when items are
added as they weren't tested before. This ensures that when throwing
`include` into the mix documents are still round-trip-able.