Add support for `include`  to `wit-smith` (#2161)
* Test fuzzers with `cargo test --workspace`

* Add `include` to `wit-smith` generation

Updates the generator to generate `include` statements in worlds to help
fuzz the `include` directive.

* Pre-sort WIT worlds at rest, update printing

This commit moves a sort that previously happened during translation to
wasm to happening during general elaboration instead, meaning that
at-rest worlds are now pre-sorted in this order. Additionally printing
of WIT worlds was updated to manually respect this order where
interfaces are printed first, then types, then functions. As a result of
this change many test expectations have been updated to shuffle items
around.

The motivation for this commit is to fix fuzz failures about
round-tripping WIT documents through wasm. With `include` now supported
more test cases are showing up which stress new paths of when items are
added as they weren't tested before. This ensures that when throwing
`include` into the mix documents are still round-trip-able.

* Re-bless some tests

* Fix test build on Windows