wasmparser(CM+GC): Allow declaring a core function type when lowering
When lowering a component function into a core function, allow the declaration
of which core function type the lowered function (or lowered CM intrinsic)
should have. Right now, this is effectively just double checking that the
declared type matches the actual lowered type. After future PRs, this will allow
declaring a Wasm GC-using function type, and lowering will adjust
accordingly. This also lets the Wasm declare which of the "equivalent" function
types it wants to lower to in the case that there are multiple rec groups
containing the "same" function type, or in the face of function type
subtyping. Again, subsequent PRs will start adding support for rec groups and
subtyping and such to CM core types.