wasmparser(CM+GC): Assert that `SubtypeCx`s are always using the same typing context (#2170)
A `SubtypeCx`, used for subtype-checking in the component model, is constructed
with two `TypesRef`s. These `TypesRef`s must be associated with the same
validator, so that we know that core types are canonicalized the same across the two of
them and we can compare `CoreTypeId`s for equality.

I had originally hoped to move the `SubtypeCx` to wrapping a single
`TypesRef`, but this cannot work because (unlike core types) new component model
types can be pushed into each different `TypesRef` for different components.