wasmparser:(CM+GC): Validate components' core function linking via canonicalization
Instead of structurally checking parameters and return types, rely on type
canonicalization to dedupe types so that we can simply compare ids. This gets
instantiation with the "same" function type defined in a different rec group to
properly fail to link.

This does not yet handle function subtyping. That will happen in follow up commits.