wasmparser(CM+GC): Validate components' core function linking via type canonicalization (#2172)
* wasmparser:(CM+GC): Validate components' core function linking via canonicalization

Instead of structurally checking parameters and return types, rely on type
canonicalization to dedupe types so that we can simply compare ids. This gets
instantiation with the "same" function type defined in a different rec group to
properly fail to link.

This does not yet handle function subtyping. That will happen in follow up commits.

* delete some unused snapshot files that got added by accident

* Add debug asserts that type lists contain IDs

* fix assertion

* Move all `desc` methods over to `Display` impls