wasmparser(CM+GC): validate `core-type` canonical option during lowering
Rather than checking that the lowered type matches the declared `core-type`
canonical option post facto, check it during the process of lowering. This sets
us up (in future commits) for peeking at the expected lowering and deciding
whether to lower to the regular linear memory version of the canonical ABI or
the GC version that I am prototyping.