Support componentizing modules that have duplicate imports. (#2145)
* Support componentizing modules that have duplicate imports.

When the `--deduplicate-imports` option is passed to `wasm-tools component new`, we rewrite the input module to compact out the duplicates. This allows tinyGo's output to be wrapped in a component. Closes #2063.

Using `Reencode` takes the maintenance pressure I feared off the deduplication engine; it won't have to chase future changes to the wasm format. Instead, `Reencode` should adapt to them and continue calling `function_index()` in all the right places.

* Typos

* Leave a slot in `dupes` even for non-function imports.

* Move integration test to wit-component, where there's infra for testing against adapters.