CM+GC: Define and validate GC lowerings for most component model types
This commit defines the expected lowerings for most component model
types (primitives, records, strings, tuples, etc..) and implements their
validation.

This commit also defines futures and streams as `externref`s in the GC ABI, but
it does not plumb that through to all of the async intrinsics yet.

Finally, this commit does *not* yet define the expected lowerings for variants
and variant-like types (options and results). This is because it isn't 100%
clear how to lower sum types to GC types yet, especially with the way that we
are doing shallow-matching to allow for the "same" type from any rec group to be
used.