wasmparser: merge VisitSimdOperator back into VisitOperator
The `simd` feature remains in place, but now does not affect the API in
this way. Rather parts of implementation pertaining to SIMD in
wasmparser get disabled, reducing the compilation time and artifact
size impact slightly, though less than in the original approach.

The results on my machine are roughly:

* origin/main simd: 4.26s producing 30M .rlib
* origin/main !simd: 3.49s producing 23M .rlib
* this simd: 4.19s producing a 26M .rlib
* this !simd: 3.96s producing a 25M .rlib

I'm not quite sure why the rlib size decreased so significantly when the
SIMD feature is enabled (I guess less inlining is occurring?) and there
is indeed a .5s hit on the compilation time to this PR.

Are those half a second worth a clunkier API? I'm not convinced :)