Support duplicate imports by default in `wasm-tools component new`
This commit is a refinement to the support added in #2145 to support
duplicate imports by default in wasm-tools, namely by avoiding changing
any binary offsets in a module except for the import section. This
implements a strategy of renaming imports in the module that goes into a
component to be unique. Classification of what an import maps to is then
ensured to use the original names, not the current names, in the module.